optimize oop
